аче)

+++ additional_functions.h
@@ -5,6 +5,6 @@ int normalize_input(char *result, int num_args, char **args);
 
 int convert_to_postfix_view(char *result, char *target, int target_length);
 
-int calculate_result(char *postfix_view);
+double calculate_result(char *postfix_view);
 
 #endif //CALC_ADDITIONAL_FUNCTIONS_H
